+++ packages/orq-rc/src/lib/config.ts
@@ -64,7 +64,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "3.3.0-rc.34",
+  sdkVersion: "3.4.0-rc.0",
   genVersion: "2.565.1",
-  userAgent: "speakeasy-sdk/typescript 3.3.0-rc.34 2.565.1 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 3.4.0-rc.0 2.565.1 2.0 @orq-ai/node",
 } as const;