+++ packages/orq-rc/RELEASES.md
@@ -1319,3 +1319,13 @@ Based on:
 - [typescript v3.3.0-rc.34] packages/orq-rc
 ### Releases
 - [NPM v3.3.0-rc.34] https://www.npmjs.com/package/@orq-ai/node/v/3.3.0-rc.34 - packages/orq-rc
+
+## 2025-04-03 20:50:49
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.528.1 (2.565.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.4.0-rc.0] packages/orq-rc
+### Releases
+- [NPM v3.4.0-rc.0] https://www.npmjs.com/package/@orq-ai/node/v/3.4.0-rc.0 - packages/orq-rc